// ApartmentApps.Client.Mobile/src/services/backend/generated/backend.generated.ts
export class UserBindingModel { 
    imageUrl: string; 
    imageThumbnailUrl: string; 
    firstName: string; 
    lastName: string; 
    unitName: string; 
    buildingName: string; 
    isTenant: boolean; 
    phoneNumber: string; 
    fullName: string; 
    address: string; 
    city: string; 
    postalCode: string; 
    email: string; 
    archived: boolean; 
    roles: string[]; 
    id: string; 
    title: string; 
    actionLinks: ActionLinkModel[];

    constructor(data?: any) {
        if (data !== undefined) {
            this.imageUrl = data["ImageUrl"] !== undefined ? data["ImageUrl"] : null;
            this.imageThumbnailUrl = data["ImageThumbnailUrl"] !== undefined ? data["ImageThumbnailUrl"] : null;
            this.firstName = data["FirstName"] !== undefined ? data["FirstName"] : null;
            this.lastName = data["LastName"] !== undefined ? data["LastName"] : null;
            this.unitName = data["UnitName"] !== undefined ? data["UnitName"] : null;
            this.buildingName = data["BuildingName"] !== undefined ? data["BuildingName"] : null;
            this.isTenant = data["IsTenant"] !== undefined ? data["IsTenant"] : null;
            this.phoneNumber = data["PhoneNumber"] !== undefined ? data["PhoneNumber"] : null;
            this.fullName = data["FullName"] !== undefined ? data["FullName"] : null;
            this.address = data["Address"] !== undefined ? data["Address"] : null;
            this.city = data["City"] !== undefined ? data["City"] : null;
            this.postalCode = data["PostalCode"] !== undefined ? data["PostalCode"] : null;
            this.email = data["Email"] !== undefined ? data["Email"] : null;
            this.archived = data["Archived"] !== undefined ? data["Archived"] : null;
            if (data["Roles"] && data["Roles"].constructor === Array) {
                this.roles = [];
                for (let item of data["Roles"])
                    this.roles.push(item);
            }
            this.id = data["Id"] !== undefined ? data["Id"] : null;
            this.title = data["Title"] !== undefined ? data["Title"] : null;
            if (data["ActionLinks"] && data["ActionLinks"].constructor === Array) {
                this.actionLinks = [];
                for (let item of data["ActionLinks"])
                    this.actionLinks.push(ActionLinkModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): UserBindingModel {
        return new UserBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["ImageUrl"] = this.imageUrl !== undefined ? this.imageUrl : null;
        data["ImageThumbnailUrl"] = this.imageThumbnailUrl !== undefined ? this.imageThumbnailUrl : null;
        data["FirstName"] = this.firstName !== undefined ? this.firstName : null;
        data["LastName"] = this.lastName !== undefined ? this.lastName : null;
        data["UnitName"] = this.unitName !== undefined ? this.unitName : null;
        data["BuildingName"] = this.buildingName !== undefined ? this.buildingName : null;
        data["IsTenant"] = this.isTenant !== undefined ? this.isTenant : null;
        data["PhoneNumber"] = this.phoneNumber !== undefined ? this.phoneNumber : null;
        data["FullName"] = this.fullName !== undefined ? this.fullName : null;
        data["Address"] = this.address !== undefined ? this.address : null;
        data["City"] = this.city !== undefined ? this.city : null;
        data["PostalCode"] = this.postalCode !== undefined ? this.postalCode : null;
        data["Email"] = this.email !== undefined ? this.email : null;
        data["Archived"] = this.archived !== undefined ? this.archived : null;
        if (this.roles && this.roles.constructor === Array) {
            data["Roles"] = [];
            for (let item of this.roles)
                data["Roles"].push(item);
        }
        data["Id"] = this.id !== undefined ? this.id : null;
        data["Title"] = this.title !== undefined ? this.title : null;
        if (this.actionLinks && this.actionLinks.constructor === Array) {
            data["ActionLinks"] = [];
            for (let item of this.actionLinks)
                data["ActionLinks"].push(item.toJS());
        }
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new UserBindingModel(JSON.parse(json));
    }
}
// ...
export class MaintenanceCheckinBindingModel { 
    statusId: string; 
    date: Date; 
    comments: string; 
    photos: ImageReference[]; 
    worker: UserBindingModel;

    constructor(data?: any) {
        if (data !== undefined) {
            this.statusId = data["StatusId"] !== undefined ? data["StatusId"] : null;
            this.date = data["Date"] ? new Date(data["Date"].toString()) : null;
            this.comments = data["Comments"] !== undefined ? data["Comments"] : null;
            if (data["Photos"] && data["Photos"].constructor === Array) {
                this.photos = [];
                for (let item of data["Photos"])
                    this.photos.push(ImageReference.fromJS(item));
            }
            this.worker = data["Worker"] ? UserBindingModel.fromJS(data["Worker"]) : null;
        }
    }

    static fromJS(data: any): MaintenanceCheckinBindingModel {
        return new MaintenanceCheckinBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["StatusId"] = this.statusId !== undefined ? this.statusId : null;
        data["Date"] = this.date ? this.date.toISOString() : null;
        data["Comments"] = this.comments !== undefined ? this.comments : null;
        if (this.photos && this.photos.constructor === Array) {
            data["Photos"] = [];
            for (let item of this.photos)
                data["Photos"].push(item.toJS());
        }
        data["Worker"] = this.worker ? this.worker.toJS() : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new MaintenanceCheckinBindingModel(JSON.parse(json));
    }
}

export class MaintenanceBindingModel { 
    userName: string; 
    userId: string; 
    name: string; 
    message: string; 
    photos: string[]; 
    unitName: string; 
    status: string; 
    scheduleDate: Date; 
    petStatus: number; 
    checkins: MaintenanceCheckinBindingModel[]; 
    user: UserBindingModel; 
    buildingName: string; 
    permissionToEnter: boolean; 
    acceptableCheckinCodes: string[]; 
    canComplete: boolean; 
    canPause: boolean; 
    canSchedule: boolean; 
    canStart: boolean;

    constructor(data?: any) {
        if (data !== undefined) {
            this.userName = data["UserName"] !== undefined ? data["UserName"] : null;
            this.userId = data["UserId"] !== undefined ? data["UserId"] : null;
            this.name = data["Name"] !== undefined ? data["Name"] : null;
            this.message = data["Message"] !== undefined ? data["Message"] : null;
            if (data["Photos"] && data["Photos"].constructor === Array) {
                this.photos = [];
                for (let item of data["Photos"])
                    this.photos.push(item);
            }
            this.unitName = data["UnitName"] !== undefined ? data["UnitName"] : null;
            this.status = data["Status"] !== undefined ? data["Status"] : null;
            this.scheduleDate = data["ScheduleDate"] ? new Date(data["ScheduleDate"].toString()) : null;
            this.petStatus = data["PetStatus"] !== undefined ? data["PetStatus"] : null;
            if (data["Checkins"] && data["Checkins"].constructor === Array) {
                this.checkins = [];
                for (let item of data["Checkins"])
                    this.checkins.push(MaintenanceCheckinBindingModel.fromJS(item));
            }
            this.user = data["User"] ? UserBindingModel.fromJS(data["User"]) : null;
            this.buildingName = data["BuildingName"] !== undefined ? data["BuildingName"] : null;
            this.permissionToEnter = data["PermissionToEnter"] !== undefined ? data["PermissionToEnter"] : null;
            if (data["AcceptableCheckinCodes"] && data["AcceptableCheckinCodes"].constructor === Array) {
                this.acceptableCheckinCodes = [];
                for (let item of data["AcceptableCheckinCodes"])
                    this.acceptableCheckinCodes.push(item);
            }
            this.canComplete = data["CanComplete"] !== undefined ? data["CanComplete"] : null;
            this.canPause = data["CanPause"] !== undefined ? data["CanPause"] : null;
            this.canSchedule = data["CanSchedule"] !== undefined ? data["CanSchedule"] : null;
            this.canStart = data["CanStart"] !== undefined ? data["CanStart"] : null;
        }
    }

    static fromJS(data: any): MaintenanceBindingModel {
        return new MaintenanceBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["UserName"] = this.userName !== undefined ? this.userName : null;
        data["UserId"] = this.userId !== undefined ? this.userId : null;
        data["Name"] = this.name !== undefined ? this.name : null;
        data["Message"] = this.message !== undefined ? this.message : null;
        if (this.photos && this.photos.constructor === Array) {
            data["Photos"] = [];
            for (let item of this.photos)
                data["Photos"].push(item);
        }
        data["UnitName"] = this.unitName !== undefined ? this.unitName : null;
        data["Status"] = this.status !== undefined ? this.status : null;
        data["ScheduleDate"] = this.scheduleDate ? this.scheduleDate.toISOString() : null;
        data["PetStatus"] = this.petStatus !== undefined ? this.petStatus : null;
        if (this.checkins && this.checkins.constructor === Array) {
            data["Checkins"] = [];
            for (let item of this.checkins)
                data["Checkins"].push(item.toJS());
        }
        data["User"] = this.user ? this.user.toJS() : null;
        data["BuildingName"] = this.buildingName !== undefined ? this.buildingName : null;
        data["PermissionToEnter"] = this.permissionToEnter !== undefined ? this.permissionToEnter : null;
        if (this.acceptableCheckinCodes && this.acceptableCheckinCodes.constructor === Array) {
            data["AcceptableCheckinCodes"] = [];
            for (let item of this.acceptableCheckinCodes)
                data["AcceptableCheckinCodes"].push(item);
        }
        data["CanComplete"] = this.canComplete !== undefined ? this.canComplete : null;
        data["CanPause"] = this.canPause !== undefined ? this.canPause : null;
        data["CanSchedule"] = this.canSchedule !== undefined ? this.canSchedule : null;
        data["CanStart"] = this.canStart !== undefined ? this.canStart : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new MaintenanceBindingModel(JSON.parse(json));
    }
}